Test cleanup/refactoring.
Remove context.py, moving (test) QApplication setup into pytest conftest.py
Added global fixture to provide handle to the QApplication instance for use in test-cases.
Removed all references to 'from context import dgp'. This was superfluous anyways as we insert the DGP module path into the system path.
