+++ tests/conftest.py
@@ -1,9 +1,14 @@
 # -*- coding: utf-8 -*-
+import os
+import sys
+import traceback
 from datetime import datetime
 from pathlib import Path
 
 import pandas as pd
 import pytest
+from PyQt5 import QtCore
+from PyQt5.QtWidgets import QApplication
 
 from dgp.core.controllers.project_controllers import AirborneProjectController
 from dgp.core.hdf5_manager import HDF5_NAME
@@ -16,12 +21,47 @@ from dgp.core.oid import OID
 from dgp.lib.gravity_ingestor import read_at1a
 from dgp.lib.trajectory_ingestor import import_trajectory
 
-# Import QApplication object for any Qt GUI test cases
-from .context import APP
+sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), '..')))
+
+"""Global pytest configuration file for DGP test suite.
+
+This takes care of configuring a QApplication instance for executing tests 
+against UI code which requires an event loop (signals etc).
+If a handle to the QApplication is required, e.g. to use as the parent to a test 
+object, the qt_app fixture can be used.
+
+The sys.excepthook is also replaced to enable catching of some critical errors
+raised within the Qt domain that would otherwise not be printed.
+
+"""
+
+
+def excepthook(type_, value, traceback_):
+    """This allows IDE to properly display unhandled exceptions which are
+    otherwise silently ignored as the application is terminated.
+    Override default excepthook with
+    >>> sys.excepthook = excepthook
+
+    See Also
+    --------
+
+    http://pyqt.sourceforge.net/Docs/PyQt5/incompatibilities.html
+    """
+    traceback.print_exception(type_, value, traceback_)
+    QtCore.qFatal('')
+
+
+sys.excepthook = excepthook
+APP = QApplication([])
 
 
 def get_ts(offset=0):
     return datetime.now().timestamp() + offset
+
+
+@pytest.fixture(scope='module')
+def qt_app():
+    return APP
 
 
 @pytest.fixture()
