+++ tests/test_controllers.py
@@ -23,7 +23,6 @@ from dgp.core.models.meter import Gravimeter
 from dgp.core.models.datafile import DataFile
 from dgp.core.controllers.flight_controller import FlightController
 from dgp.core.models.flight import Flight
-from .context import APP
 
 
 def test_attribute_proxy(tmpdir):
