+++ tests/test_etc.py
@@ -1,4 +1,3 @@
-from .context import dgp
 import unittest
 import numpy as np
 import pandas as pd
