+++ tests/test_loader.py
@@ -7,7 +7,6 @@ import pytest
 from PyQt5.QtTest import QSignalSpy
 from pandas import DataFrame
 
-from .context import APP
 
 from dgp.core.file_loader import FileLoader
 
@@ -23,8 +22,8 @@ def mock_failing_loader(*args, **kwargs):
     raise FileNotFoundError
 
 
-def test_load_mock():
-    loader = FileLoader(Path(TEST_FILE_GRAV), mock_loader, APP)
+def test_load_mock(qt_app):
+    loader = FileLoader(Path(TEST_FILE_GRAV), mock_loader, qt_app)
     spy_complete = QSignalSpy(loader.loaded)
     spy_error = QSignalSpy(loader.error)
 
@@ -38,7 +37,7 @@ def test_load_mock():
     assert 0 == len(spy_error)
 
 
-def test_load_failure():
+def test_load_failure(qt_app):
     called = False
 
     def _error_handler(exception: Exception):
@@ -46,7 +45,7 @@ def test_load_failure():
         nonlocal called
         called = True
 
-    loader = FileLoader(Path(), mock_failing_loader, APP)
+    loader = FileLoader(Path(), mock_failing_loader, qt_app)
     loader.error.connect(_error_handler)
     spy_err = QSignalSpy(loader.error)
     assert 0 == len(spy_err)
